backend/pyproject.toml
[project]
name = "ai-study-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "firecrawl-py",
    "gradio",
    "langchain",
    "langchain-community",
    "langchain-openai",
    "langgraph",
    "langfuse",
    "pydantic",
    "pypdf",
    "python-docx",
    "python-dotenv",
    "python-magic",
    "tavily-python",
    "tiktoken",
    "unstructured",
    "fastapi>=0.128.7",
    "uvicorn>=0.40.0",
    "python-multipart>=0.0.22",
]
